+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make `Safety Code Length` option an enum instead of integer.
 - Capitalize protocol field names.
 - Show Safety Code field only if length > 0.
+- Show payload length instead of raw bytes for more clarity.
 
 ### Fixed